prepend postcss ONLY for rolled up indiv. comps.

+++ svelte.config.js
@@ -1,6 +1,8 @@
 // import adapter from '@sveltejs/adapter-auto';
 import adapter from '@sveltejs/adapter-static';
 import sveltePreprocess from 'svelte-preprocess';
+import preprocessConfig from './svelte.preprocess.config.js';
+
 import { mdsvex } from 'mdsvex';
 
 import path, { dirname } from 'path';
@@ -18,20 +20,20 @@ const __dirname = dirname(__filename);
 // -cw compile and watch for file changes during development
 const args = process.env.NODE_ENV === 'development' ? '-cw' : '-c';
 
-// if (true && process.env.NODE_ENV !== 'development') {
-// rollup the components tocompiled code into /dist/ folder
-let server = spawn('rollup', [args], {
-	stdio: ['ignore', 'inherit', 'inherit'],
-	shell: true
-});
+if (true || process.env.NODE_ENV !== 'development') {
+	// rollup the components tocompiled code into /dist/ folder
+	let server = spawn('rollup', [args], {
+		stdio: ['ignore', 'inherit', 'inherit'],
+		shell: true
+	});
 
-process.on('SIGTERM', () => {
-	if (server) server.kill(0);
-});
-process.on('exit', () => {
-	if (server) server.kill(0);
-});
-// }
+	process.on('SIGTERM', () => {
+		if (server) server.kill(0);
+	});
+	process.on('exit', () => {
+		if (server) server.kill(0);
+	});
+}
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -39,13 +41,7 @@ const config = {
 	preprocess: [
 		sveltePreprocess({
 			sourceMap: process.env.NODE_ENV !== 'development',
-			// scss: true,
-			// postcss: true, // set postcss: true if postcss-load-config is installed and svelte-preprocess will look for a PostCSS config file in your project.,
-			postcss: {
-				configFilePath: path.resolve(__dirname, './postcss.config.js'),
-				prependData: `@import '${path.resolve('./src/lib/app.css')}';`
-			},
-			globalStyle: {} // enables us to have :global css
+			postcss: true // set postcss: true if postcss-load-config is installed and svelte-preprocess will look for a PostCSS config file in your project.,
 		}),
 		mdsvex({
 			extensions: ['.svx', '.md', '.svelte.md'],
